Segregate shard level tests for node to node and remote store segment replication
